refactor: Move wasm helper to node/utils
