+++ src/node/server/serverPluginWasm.ts
@@ -1,6 +1,6 @@
 import { ServerPlugin } from '.'
 import { isImportRequest } from '../utils'
-import * as wasmHelper from '../../client/wasmHelper'
+import * as wasmHelper from '../utils/wasmHelper'
 
 export const wasmPlugin: ServerPlugin = ({ app }) => {
   app.use((ctx, next) => {
