+++ src/node/utils/wasmHelper.ts
